Move napari-workflows input models to task module
