+++ tests/tasks/test_workflows_napari_workflows.py
@@ -25,10 +25,14 @@ from ._validation import validate_axes_and_coordinateTransformations
 from ._validation import validate_labels_and_measurements
 from ._validation import validate_schema
 from .lib_empty_ROI_table import _add_empty_ROI_table
-from fractal_tasks_core.ome_zarr.input_models import NapariWorkflowsInput
-from fractal_tasks_core.ome_zarr.input_models import NapariWorkflowsOutput
 from fractal_tasks_core.tasks.napari_workflows_wrapper import (
     napari_workflows_wrapper,
+)
+from fractal_tasks_core.tasks.napari_workflows_wrapper import (
+    NapariWorkflowsInput,
+)
+from fractal_tasks_core.tasks.napari_workflows_wrapper import (
+    NapariWorkflowsOutput,
 )
 from fractal_tasks_core.zarr import OverwriteNotAllowedError
